+++ scripts/callgraph2pdf.sh
@@ -11,12 +11,16 @@ fi
 
 if [ -f "$jsonfile" ]
 then
-    dotfile=${jsonfile%%.json}.dot
-    pdffile=${jsonfile%%.json}.pdf
-    $scriptDir/gprof2dot.py -f json $jsonfile > $dotfile
-    dot -Tpdf $dotfile -o $pdffile
-    echo "[SUCCESS] Generated $pdffile"
+    dotfileTop=${jsonfile%%.json}Top.dot
+    dotfileAll=${jsonfile%%.json}All.dot
+    pdffileTop=${jsonfile%%.json}Top.pdf
+    pdffileAll=${jsonfile%%.json}All.pdf
+    $scriptDir/gprof2dot.py -f json $jsonfile > $dotfileTop
+    # --skew=0.01 
+    $scriptDir/gprof2dot.py -f json $jsonfile -e0 -n0 > $dotfileAll
+    dot -Tpdf $dotfileTop -o $pdffileTop
+    dot -Tpdf $dotfileAll -o $pdffileAll
+    echo "[SUCCESS] Generated $pdffileTop and $pdffileAll"
 else
     echo "[ERROR] $jsonfile not found."
 fi
-